+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 10.06.2018
 
+* [Rust] Code Royale - Building only one giant (CodinGame)
 * [Rust] Code Royale - Refactoring (CodinGame)
 
 #### 9.06.2018